requirements.txt
appdirs==1.4.4
attrs==23.2.0
Babel==2.10.3
bcc==0.29.1
bcrypt==3.2.2
beautifulsoup4==4.12.3
blinker==1.7.0
bottle==0.12.25
Brlapi==0.8.5
Brotli==1.1.0
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
cloud-init==25.1.4
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
contourpy==1.0.7
cryptography==41.0.7
cssselect==1.2.0
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.3.2
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.9.0
distro-info==1.7+build1
docker==5.0.3
duplicity==2.1.4
fasteners==0.18
fonttools==4.46.0
fs==2.4.16
Glances==3.4.0.3
greenlet==3.0.3
html5lib==1.1
httplib2==0.20.4
idna==3.6
influxdb==5.3.1
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
kiwisolver==0.0.0
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.29.0
lxml==5.2.1
lz4==4.0.2+dfsg
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.6.3
mdurl==0.1.2
monotonic==1.6
mpmath==0.0.0
msgpack==1.0.3
netaddr==0.8.0
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.2
olefile==0.46
packaging==24.0
paramiko==2.12.0
pexpect==4.9.0
pillow==10.2.0
ply==3.11
psutil==5.9.8
ptyprocess==0.7.0
pyasn1==0.4.8
pycairo==1.25.1
pycryptodomex==3.20.0
pycups==2.0.1
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.7.0
PyNaCl==1.5.0
pynvim==0.5.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
pysmi==0.3.4
pysnmp==4.4.12
pystache==0.6.0
python-apt==2.7.7+ubuntu4
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
SciPy==1.11.4
setuptools==68.1.2
six==1.16.0
soupsieve==2.5
sympy==1.12
systemd-python==235
typing_extensions==4.10.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.16.0
ufw==0.36.2
ujson==5.9.0
unattended-upgrades==0.1
unicodedata2==15.1.0
urllib3==2.0.7
wadllib==1.3.6
webencodings==0.5.1
websocket-client==1.7.0
wheel==0.42.0
xdg==5
xkit==0.0.0
